backend/requirements.txt
bleak==0.22.3
fastapi==0.115.8
influxdb_client==1.48.0
python-dotenv==1.0.1
redis==5.2.1
uvicorn==0.34.0
uvicorn[standard]